WIP vm

+++ kernel/c/x86.h
@@ -64,9 +64,9 @@ lgdt(struct segdesc *p, int size)
 {
   volatile ushort pd[3];
 
-  pd[0] = size-1;
-  pd[1] = (uint)p;
-  pd[2] = (uint)p >> 16;
+  pd[0] = (ushort)(size-1);
+  pd[1] = (ushort)(uint)p;
+  pd[2] = (ushort)((uint)p >> 16);
 
   asm volatile("lgdt (%0)" : : "r" (pd));
 }
@@ -78,9 +78,9 @@ lidt(struct gatedesc *p, int size)
 {
   volatile ushort pd[3];
 
-  pd[0] = size-1;
-  pd[1] = (uint)p;
-  pd[2] = (uint)p >> 16;
+  pd[0] = (ushort)(size-1);
+  pd[1] = (ushort)(uint)p;
+  pd[2] = (ushort)((uint)p >> 16);
 
   asm volatile("lidt (%0)" : : "r" (pd));
 }
